Style google places auto-complete field
