+++ components/newListing/Address/AutoComplete/styles.js
@@ -5,7 +5,7 @@ import * as colors from '@/assets/colors'
 export default StyleSheet.create({
   container: {
     display: 'flex',
-    flex: 1
+    flex: 0
   },
   textInputContainer: {
     display: 'flex',
@@ -20,5 +20,13 @@ export default StyleSheet.create({
     fontSize: 18,
     lineHeight: Platform.OS === 'ios' ? null : 18,
     color: colors.gray.dark
+  },
+  listView: {
+    borderWidth: 1,
+    borderColor: colors.blue.medium,
+    borderBottomLeftRadius: 4,
+    borderBottomRightRadius: 4,
+    marginTop: -3,
+    backgroundColor: 'white'
   }
 })
